chore: move account validation into shared file
